+++ test/CircuitProofGeneration.test.ts
@@ -5,13 +5,15 @@ import { CircuitZKitConfig } from "@solarity/zkit";
 
 import { generateAST } from "./helpers/generator";
 
-import CircuitTypesGenerator from "../src/core/CircuitTypesGenerator";
+import { CircuitTypesGenerator } from "../src";
+import { findProjectRoot } from "../src/utils";
 
 describe("Circuit Proof Generation", function () {
   const astDir = "test/cache/circuits-ast";
 
   const circuitTypesGenerator = new CircuitTypesGenerator({
     basePath: "test/fixture",
+    projectRoot: findProjectRoot(process.cwd()),
     circuitsASTPaths: [
       "test/cache/circuits-ast/Basic.json",
       "test/cache/circuits-ast/credentialAtomicQueryMTPV2OnChainVoting.json",
@@ -49,8 +51,11 @@ describe("Circuit Proof Generation", function () {
     new (await circuitTypesGenerator.getCircuitObject("CredentialAtomicQueryMTPOnChainVoting"))();
     new (await circuitTypesGenerator.getCircuitObject("EnhancedMultiplier"))();
 
-    await expect(circuitTypesGenerator.getCircuitObject("Multiplier2")).to.be.rejectedWith(
-      "Circuit Multiplier2 type does not exist.",
+    await expect(circuitTypesGenerator.getCircuitObject("Multiplier3")).to.be.rejectedWith(
+      "Circuit Multiplier3 type does not exist.",
+    );
+    await expect(circuitTypesGenerator.getCircuitObject("test/fixture/Basic.circom:Multiplier3")).to.be.rejectedWith(
+      "Circuit Multiplier3 type does not exist.",
     );
   });