+++ test/CircuitTypesGenerator.test.ts
@@ -7,7 +7,7 @@ import { generateAST } from "./helpers/generator";
 
 import { findProjectRoot } from "../src/utils";
 
-import CircuitTypesGenerator from "../src/core/CircuitTypesGenerator";
+import { CircuitTypesGenerator } from "../src";
 
 describe("Circuit Types Generation", function () {
   const expectedTypes = ["core/CredentialAtomicQueryMTPOnChainVoting.ts", "core/Multiplier2.ts"];
@@ -18,6 +18,7 @@ describe("Circuit Types Generation", function () {
 
   const circuitTypesGenerator = new CircuitTypesGenerator({
     basePath: "test/fixture",
+    projectRoot: findProjectRoot(process.cwd()),
     circuitsASTPaths: [
       "test/cache/circuits-ast/Basic.json",
       "test/cache/circuits-ast/credentialAtomicQueryMTPV2OnChainVoting.json",